rename package name - invest_amount to amount
