+++ src/main/java/application/delegator/InvestmentAmountReaderDelegator.java
@@ -3,7 +3,7 @@ package application.delegator;
 import java.io.BufferedReader;
 import java.io.IOException;
 
-import domain.invest_amount.InvestmentAmount;
+import domain.amount.InvestmentAmount;
 import domain.type.InvestmentType;
 
 public interface InvestmentAmountReaderDelegator {
