+++ src/main/java/application/delegator/RegistryBasedInvestmentAmountDelegator.java
@@ -4,7 +4,7 @@ import java.io.BufferedReader;
 import java.io.IOException;
 
 import application.registry.InvestmentAmountReaderRegistry;
-import domain.invest_amount.InvestmentAmount;
+import domain.amount.InvestmentAmount;
 import domain.type.InvestmentType;
 
 public class RegistryBasedInvestmentAmountDelegator implements InvestmentAmountReaderDelegator {
