+++ src/main/java/application/factory/DefaultInvestmentFactory.java
@@ -16,9 +16,9 @@ import application.resolver.KoreanStringBasedTaxableResolver;
 import application.resolver.TaxableResolver;
 import domain.interest_rate.AnnualInterestRate;
 import domain.interest_rate.InterestRate;
-import domain.invest_amount.FixedDepositAmount;
-import domain.invest_amount.InstallmentInvestmentAmount;
-import domain.invest_amount.LumpSumInvestmentAmount;
+import domain.amount.FixedDepositAmount;
+import domain.amount.InstallmentInvestmentAmount;
+import domain.amount.LumpSumInvestmentAmount;
 import domain.invest_period.InvestPeriod;
 import domain.invest_period.MonthBasedRemainingPeriodProvider;
 import domain.invest_period.PeriodMonthsRange;
