+++ src/main/java/application/parser/FixedDepositInvestmentAmountParser.java
@@ -1,7 +1,7 @@
 package application.parser;
 
-import domain.invest_amount.FixedDepositAmount;
-import domain.invest_amount.InvestmentAmount;
+import domain.amount.FixedDepositAmount;
+import domain.amount.InvestmentAmount;
 
 public class FixedDepositInvestmentAmountParser implements InvestmentAmountParser {
 	@Override
