+++ src/main/java/application/parser/InvestmentAmountParser.java
@@ -1,6 +1,6 @@
 package application.parser;
 
-import domain.invest_amount.InvestmentAmount;
+import domain.amount.InvestmentAmount;
 
 public interface InvestmentAmountParser {
 	InvestmentAmount parse(String line);
