+++ src/main/java/application/reader/InvestmentAmountReader.java
@@ -3,7 +3,7 @@ package application.reader;
 import java.io.BufferedReader;
 import java.io.IOException;
 
-import domain.invest_amount.InvestmentAmount;
+import domain.amount.InvestmentAmount;
 import domain.type.InvestmentType;
 
 public interface InvestmentAmountReader {
