+++ src/main/java/domain/amount/InstallmentInvestmentAmount.java
@@ -1,0 +1,5 @@
+package domain.amount;
+
+public interface InstallmentInvestmentAmount extends InvestmentAmount {
+	int getMonthlyAmount();
+}
