+++ src/main/java/domain/amount/LumpSumInvestmentAmount.java
@@ -1,9 +1,9 @@
-package domain.invest_amount;
+package domain.amount;
 
 /**
  * 예치 금액 관리하는 인터페이스
  */
-public interface LumpSumInvestmentAmount extends InvestmentAmount{
+public interface LumpSumInvestmentAmount extends InvestmentAmount {
 	int getDepositAmount();
 
 	int calCompoundInterest(double totalGrowthFactor);
