+++ src/main/java/domain/amount/TargetAmountReachable.java
@@ -1,4 +1,4 @@
-package domain.invest_amount;
+package domain.amount;
 
 public interface TargetAmountReachable {
