--- src/main/java/domain/invest_amount/InstallmentInvestmentAmount.java
@@ -1,5 +1,0 @@
-package domain.invest_amount;
-
-public interface InstallmentInvestmentAmount extends InvestmentAmount{
-	int getMonthlyAmount();
-}
