+++ src/main/java/domain/invest_period/InvestPeriod.java
@@ -1,6 +1,6 @@
 package domain.invest_period;
 
-import domain.invest_amount.InstallmentInvestmentAmount;
+import domain.amount.InstallmentInvestmentAmount;
 
 public interface InvestPeriod {
 	int getMonths();
